Extract ExhibitsController behaviors into concerns
This enables our local application to customize the behavior of
Spotlight::ExhibitsController.

Our plan is to save an associated user editable object (ExhibitFilter)
in the create and update actions.
